+++ examples/isolines.lua
@@ -2,7 +2,7 @@
 -- Here we generate a pattern randomly filled with points, and take as a scalar
 -- field `N(cell) = F_2(cell) - F_1(cell)`, where `F_n` is the Chebyshev distance
 -- to the nth nearest neighbour. Isolines at `N = 0` are drawn by thresholding `N`
--- at 1 and taking the surface.
+-- at 1 and taking the interior hull.
 
 local cell          = require('forma.cell')
 local subpattern    = require('forma.subpattern')
@@ -26,7 +26,7 @@ local mask = function(tcell)
     return F2 - F1  > 1
 end
 
--- Compute the thresholded pattern and print its surface
+-- Compute the thresholded pattern and print its interior hull
 local noise = subpattern.mask(sq, mask)
-subpattern.print_patterns(sq, {noise:surface()}, {'#'})
+subpattern.print_patterns(sq, {noise:interior_hull()}, {'#'})